requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
Pillow>=11.0.0
stripe==10.12.0
python-dotenv==1.0.1
requests==2.32.3